Add a `MacOsNotarizationCredentials` variant to use a keychain profile for notarytool settings and password. (#353)
* Add a `MacOsNotarizationCredentials` variant to use a keychain profile for notarytool settings and password.

This allows notarytool to use an Apple ID, Team ID and app-specific password stored in the keychain, by just providing the keychain profile as an env var. This reduces the number of env vars needed, and means the password isn't in a variable as plain text.

This approach is described under "Upload your app to the notarization service" here: https://developer.apple.com/documentation/security/customizing-the-notarization-workflow

* Apply suggestions from code review

* add change file

---------

Co-authored-by: Lucas Nogueira <118899497+lucasfernog-crabnebula@users.noreply.github.com>