packager: codesign: macos: Fix certificates
Currently, it is not possible to pass signing_certificate and
signing_certificate_password using  Cargo.toml of config json. Instead,
from what I gather, environment variables should be used for these
variables. However, the current logic to fallback to environment variables
is erroneous and will only use environment variables if no macos property
is defined, as opposed to if the variable is not found in the config.

Signed-off-by: Ayush Singh <ayush@beagleboard.org>